[Part-1] Refactor k8s related code into a separate module
This commit is first of the chain in refactoring the cluster/
package.
