+++ .gitignore
@@ -12,3 +12,4 @@ node_modules
 .DS_Store
 ui-tests/playwright-report
 ui-tests/test-results
+/src/codeflare_sdk.egg-info/
