+++ src/codeflare_sdk/common/kubernetes_cluster/auth.py
@@ -23,7 +23,7 @@ import abc
 from kubernetes import client, config
 import os
 import urllib3
-from ..utils.kube_api_helpers import _kube_api_error_handling
+from .kube_api_helpers import _kube_api_error_handling
 
 from typing import Optional
